Add tests for Aegis JSON importer
This adds tests for `JsonImporter.OpenJsonAegis`, so that I could test
my changes without running the whole application and so that the quality
of the code is easier to maintain over time.

Since the returned structure is quite complex and might be temporary, I
decided against using plain asserts and used the Verify library instead.
It's a bit sophisticated but it does the asserting stuff automatically
(and shows diffs of differences), which saves a lot of time.

Use `dotnet test` to check the tests. On Linux, use `dotnet test
AuthDesk.CoreTests` (since the other projects cannot be built).

To update review and update test snapshots automatically, you can use
the Verify CLI (`dotnet-verify review`). There may be an integration
with Visual Studio but I haven't tried it.